+++ index.js
@@ -11,6 +11,7 @@ module.exports = {
     'no-mutating-assign': require('./rules/no-mutating-assign'),
     'no-mutation': require('./rules/no-mutation'),
     'no-nil': require('./rules/no-nil'),
+    'no-proxy': require('./rules/no-proxy'),
     'no-this': require('./rules/no-this'),
     'no-throw': require('./rules/no-throw'),
     'no-unused-expression': require('./rules/no-unused-expression')
@@ -28,6 +29,7 @@ module.exports = {
         'fp/no-mutating-assign': 'error',
         'fp/no-mutation': 'error',
         'fp/no-nil': 'error',
+        'fp/no-proxy': 'error',
         'fp/no-this': 'error',
         'fp/no-throw': 'error',
         'fp/no-unused-expression': 'error'